Cleaned up a bunch more old references
