+++ tap-agent/examples/payment_flow.rs
@@ -12,11 +12,11 @@ use std::collections::HashMap;
 use std::str::FromStr;
 use std::sync::Arc;
 
-use didcomm::secrets::{Secret, SecretMaterial, SecretType};
 use tap_agent::agent::{Agent, DefaultAgent};
 use tap_agent::config::AgentConfig;
 use tap_agent::crypto::{BasicSecretResolver, DefaultMessagePacker};
 use tap_agent::did::{KeyResolver, MultiResolver};
+use tap_agent::key_manager::{Secret, SecretMaterial, SecretType};
 use tap_caip::AssetId;
 use tap_msg::message::types::{Authorize, Settle};
 use tap_msg::{Participant, Payment};
