+++ tap-agent/examples/transfer_flow.rs
@@ -11,12 +11,12 @@ use std::collections::HashMap;
 use std::str::FromStr;
 use std::sync::Arc;
 
-use tap_msg::didcomm::{Secret, SecretMaterial, SecretType};
 use tap_agent::agent::{Agent, DefaultAgent};
 use tap_agent::config::AgentConfig;
 use tap_agent::crypto::{BasicSecretResolver, DefaultMessagePacker};
 use tap_agent::did::{KeyResolver, MultiResolver};
 use tap_caip::AssetId;
+use tap_msg::key_manager::{Secret, SecretMaterial, SecretType};
 use tap_msg::message::{Authorize, Settle, Transfer};
 use tap_msg::Participant;
