+++ tap-agent/src/did.rs
@@ -7,6 +7,7 @@
 //!
 //! It also provides functionality to generate new DIDs with different cryptographic curves.
 
+use crate::key_manager::{Secret, SecretMaterial, SecretType};
 use async_trait::async_trait;
 use base64::Engine;
 use curve25519_dalek::edwards::CompressedEdwardsY;
@@ -15,16 +16,108 @@ use k256::ecdsa::SigningKey as Secp256k1SigningKey;
 use multibase::{decode, encode, Base};
 use p256::ecdsa::SigningKey as P256SigningKey;
 use rand::rngs::OsRng;
-use tap_msg::didcomm::{
-    DIDDoc, Secret, SecretMaterial, SecretType, VerificationMaterial, VerificationMethod,
-    VerificationMethodType,
-};
-
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
 use std::collections::HashMap;
 use std::fmt::Debug;
 use std::sync::{Arc, RwLock};
 
 use crate::error::{Error, Result};
+
+/// DID Document
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct DIDDoc {
+    /// DID that this document describes
+    pub id: String,
+
+    /// List of verification methods
+    pub verification_method: Vec<VerificationMethod>,
+
+    /// List of authentication verification method references (id strings)
+    pub authentication: Vec<String>,
+
+    /// List of key agreement verification method references (id strings)
+    pub key_agreement: Vec<String>,
+
+    /// List of services
+    pub service: Vec<Service>,
+}
+
+/// Service definition in a DID Document
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct Service {
+    /// Service ID
+    pub id: String,
+
+    /// Service type
+    #[serde(rename = "type")]
+    pub type_: String,
+
+    /// Service endpoint URL
+    pub service_endpoint: String,
+
+    /// Additional properties
+    #[serde(flatten)]
+    pub properties: HashMap<String, Value>,
+}
+
+/// Verification method in a DID Document
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct VerificationMethod {
+    /// Verification method ID
+    pub id: String,
+
+    /// Verification method type
+    #[serde(rename = "type")]
+    pub type_: VerificationMethodType,
+
+    /// Controller DID
+    pub controller: String,
+
+    /// Verification material
+    #[serde(flatten)]
+    pub verification_material: VerificationMaterial,
+}
+
+/// Verification method type
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub enum VerificationMethodType {
+    /// Ed25519 Verification Key 2018
+    Ed25519VerificationKey2018,
+
+    /// X25519 Key Agreement Key 2019
+    X25519KeyAgreementKey2019,
+
+    /// ECDSA Secp256k1 Verification Key 2019
+    EcdsaSecp256k1VerificationKey2019,
+
+    /// JSON Web Key 2020
+    JsonWebKey2020,
+}
+
+/// Verification material
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[serde(untagged)]
+pub enum VerificationMaterial {
+    /// Base58 encoded public key
+    Base58 {
+        /// Public key encoded in base58
+        public_key_base58: String,
+    },
+
+    /// Multibase encoded public key
+    Multibase {
+        /// Public key encoded in multibase
+        public_key_multibase: String,
+    },
+
+    /// JSON Web Key
+    JWK {
+        /// Public key in JWK format
+        public_key_jwk: Value,
+    },
+}
 
 /// Key types supported for DID generation
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
