+++ tap-http/src/main.rs
@@ -8,9 +8,9 @@ use std::path::PathBuf;
 use std::process;
 use std::sync::Arc;
 use tap_agent::DefaultAgent;
-use tap_msg::didcomm::{Secret, SecretMaterial, SecretType};
 use tap_http::event::{EventLoggerConfig, LogDestination};
 use tap_http::{TapHttpConfig, TapHttpServer};
+use tap_msg::key_manager::{Secret, SecretMaterial, SecretType};
 use tap_node::{NodeConfig, TapNode};
 use tracing::{debug, error, info};
 
@@ -218,9 +218,7 @@ fn create_agent(
                         Secret {
                             type_: SecretType::JsonWebKey2020,
                             id: format!("{}#keys-1", did),
-                            secret_material: SecretMaterial::JWK {
-                                private_key_jwk,
-                            },
+                            secret_material: SecretMaterial::JWK { private_key_jwk },
                         }
                     } else {
                         return Err(format!(
