+++ tap-node/tests/message_tests.rs
@@ -1,15 +1,15 @@
 //! Tests for message handling in TAP Node
 
 use async_trait::async_trait;
-use didcomm::did::{DIDDoc, VerificationMaterial, VerificationMethod, VerificationMethodType};
 use std::collections::HashMap;
 use std::fmt::Debug;
 use std::sync::Arc;
 use tap_agent::crypto::DebugSecretsResolver;
 use tap_agent::did::SyncDIDResolver;
+use tap_agent::did::{DIDDoc, VerificationMaterial, VerificationMethod, VerificationMethodType};
 use tap_agent::error::Result;
 use tap_agent::{AgentConfig, DefaultAgent};
-use tap_msg::didcomm::Message;
+use tap_msg::didcomm::PlainMessage;
 use tap_msg::message::{ErrorBody, TapMessageBody};
 use tap_node::NodeConfig;
