+++ tap-node/tests/validation_tests.rs
@@ -4,7 +4,7 @@
 //! specifically testing the ValidationMessageProcessor implementation.
 
 use serde_json::json;
-use tap_msg::didcomm::Message;
+use tap_msg::didcomm::PlainMessage;
 use tap_node::message::processor::{MessageProcessor, ValidationMessageProcessor};
 
 #[tokio::test]
